pyproject.toml
[project]
name = "mcp-practical"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "fastmcp>=0.1.0",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "mcp[cli]==1.6.0",
    "nest-asyncio>=1.6.0",
    "openai==1.75.0",
    "python-dotenv>=1.0.0",
]
